# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Spring Security JWT Audit API",
    "description" : "REST API secured with JWT: a short-lived access token in the `Authorization` header and a long-lived refresh token in an `HttpOnly` cookie, rotated on every use.\n\n**Authenticating in this page**\n1. Call `POST /api/v1/auth/login` with one of the demo accounts (`admin` / `manager` / `user`, password `Demo1234!`).\n2. Copy `authToken` from the response.\n3. Click **Authorize** and paste it. The token is valid for 15 minutes.\n\nThe refresh cookie is set by the browser and scoped to `/api/v1/auth`, so `POST /api/v1/auth/refresh` works from this page without pasting anything.\n\n**Roles** are hierarchical: `SYSTEM` > `ADMIN` > `MANAGER` > `USER`. A role inherits every permission below it.\n\nAuthentication endpoints are rate limited per client IP and answer `429` with a `Retry-After` header once the limit is hit.",
    "contact" : {
      "name" : "PaulStna",
      "url" : "https://github.com/PaulStna"
    },
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "This server"
  } ],
...
  "paths" : {
    "/api/v1/users/{id}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get one user",
        "description" : "Requires MANAGER.",
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The user.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No user with that id.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No access token, or it is invalid or expired.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Authenticated, but the role is not enough.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected failure. The body carries a traceId that matches the log entry; no internal detail is exposed.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Users" ],
        "summary" : "Update a user",
        "description" : "Requires MANAGER. Omit the password to keep the current one. The same ceiling on grantable roles applies as on create.",
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "User updated.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "The payload breaks a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No user with that id, or an unknown role.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The new username is already taken.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No access token, or it is invalid or expired.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Authenticated, but the role is not enough.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected failure. The body carries a traceId that matches the log entry; no internal detail is exposed.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Users" ],
        "summary" : "Delete a user",
        "description" : "Requires ADMIN. Deletes the account and its sessions. Roles are a shared catalogue and are left untouched.",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "User deleted."
          },
          "404" : {
            "description" : "No user with that id.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "No access token, or it is invalid or expired.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Authenticated, but the role is not enough.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected failure. The body carries a traceId that matches the log entry; no internal detail is exposed.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UserRequest" : {
        "type" : "object",
        "description" : "Payload for creating or updating a user. The password is required on create and optional on update, where omitting it keeps the current one.",
        "properties" : {
          "username" : {
            "type" : "string",
            "description" : "Username may only contain letters, digits, dots, underscores and hyphens",
            "example" : "ada.lovelace",
            "maxLength" : 50,
            "minLength" : 3,
            "pattern" : "^[a-zA-Z0-9._-]+$"
          },
          "password" : {
            "type" : "string",
            "format" : "password",
            "description" : "Password must contain an uppercase letter, a lowercase letter, a digit and a special character. Omit on update to keep the current one.",
            "example" : "Str0ng!Pass",
            "maxLength" : 72,
            "minLength" : 8,
            "pattern" : "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[^a-zA-Z0-9]).+$"
          },
          "roles" : {
            "type" : "array",
            "description" : "Role names, with or without the ROLE_ prefix. Only roles at or below the caller's own level may be granted, so SYSTEM is never assignable. Omit for a plain USER.",
            "example" : [ "MANAGER" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          }
        },
        "required" : [ "username" ]
      },
      "ErrorResponse" : {
        "description" : "Every failure in this API has this shape.",
        "properties" : {
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 400
          },
          "message" : {
            "type" : "string",
            "description" : "Safe to show a user. Never carries a stack trace or internal detail.",
            "example" : "Validation failed"
          },
          "path" : {
            "type" : "string",
            "example" : "/api/v1/users"
          },
          "traceId" : {
            "type" : "string",
            "description" : "The id this request was logged under. Quote it and the exact request can be found in the logs.",
            "example" : "b8f1c2d3-4e5a-6b7c-8d9e-0f1a2b3c4d5e"
          },
          "fieldErrors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string",
              "description" : "Only present on a validation failure.",
              "example" : "{\"password\":\"Password must be between 8 and 72 characters\"}"
            },
            "description" : "Only present on a validation failure.",
            "example" : {
              "password" : "Password must be between 8 and 72 characters"
            }
          }
        }
      },
      "UserResponse" : {
        "type" : "object",
        "description" : "A user as the API exposes it. Deliberately not the entity: the password hash and the session tokens have no field here to leak through.",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "6c2f1b7e-6f1a-4c3e-9f4a-2b8d5c1e7a90"
          },
          "username" : {
            "type" : "string",
            "example" : "ada.lovelace"
          },
          "roles" : {
            "type" : "array",
            "description" : "Granted roles, always prefixed with ROLE_.",
            "example" : [ "ROLE_MANAGER" ],
            "items" : {
              "type" : "string"
            },
            "uniqueItems" : true
          },
          "enabled" : {
            "type" : "boolean",
            "description" : "A disabled account cannot sign in, and its live tokens stop working."
          },
          "accountNonLocked" : {
            "type" : "boolean",
            "description" : "False once the account is locked. Same effect as disabling, different reason in the audit trail."
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Access token returned by `/auth/login`. Paste the raw token; the `Bearer ` prefix is added for you.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "refreshCookie" : {
        "type" : "apiKey",
        "description" : "Set by `/auth/login` and `/auth/register`. `HttpOnly`, so it cannot be read from JavaScript, and sent only to `/api/v1/auth`.",
        "name" : "refreshToken",
        "in" : "cookie"
      }
    }
  }
}